requirements.txt
# Deep Learning
# Note: TensorFlow 2.16+ requires Keras 3.x
# For Apple Silicon (M4 Pro) MPS support, TensorFlow >= 2.10.0 is required
# Keras is automatically installed with TensorFlow in the correct version
tensorflow>=2.10.0
tf-models-official

# Image Processing
opencv-python>=4.6.0,<5.0.0
Pillow>=9.0.0,<11.0.0
albumentations>=1.3.0,<2.0.0

# Data Science
numpy>=1.23.0,<2.0.0
pandas>=1.5.0,<3.0.0
matplotlib>=3.6.0,<4.0.0
seaborn>=0.12.0,<1.0.0
mlflow==2.8.0
optuna>=3.0.0,<4.0.0

# API
fastapi>=0.95.0,<1.0.0
uvicorn[standard]>=0.20.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
pydantic>=1.10.0,<3.0.0

# Web App (choose one or both)
streamlit>=1.25.0,<2.0.0
# flask>=2.3.0,<4.0.0

# Utilities
tqdm>=4.65.0,<5.0.0
scikit-learn>=1.2.0,<2.0.0

# Jupyter
jupyter>=1.0.0,<2.0.0
ipykernel>=6.20.0,<7.0.0
notebook>=6.5.0,<8.0.0

# Azure (optional, for deployment)
azure-storage-blob>=12.14.0,<13.0.0
azure-identity>=1.12.0,<2.0.0

# Additional utilities
scipy>=1.9.0,<2.0.0
imageio>=2.25.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Testing
pytest>=7.0.0,<9.0.0
